+++ aws/s3/main.tf
@@ -87,7 +87,7 @@ data "aws_iam_policy_document" "this" {
   }
 
   dynamic "statement" {
-    for_each = var.enabled_s3_http_access ? [1] : []
+    for_each = var.disabled_s3_http_access ? [1] : []
     content {
       actions = [
         "s3:*",