+++ aws/s3/variables.tf
@@ -198,7 +198,7 @@ variable "write_access_logs_source_bucket_arns" {
   default     = []
 }
 
-variable "enabled_s3_http_access" {
+variable "disabled_s3_http_access" {
   description = "Whether to restrict HTTP access to S3 bucket."
   type        = bool
   default     = true