+++ aws/cloudtrail/variables.tf
@@ -119,3 +119,8 @@ variable "versioning_status" {
   default     = "Disabled"
 }
 
+variable "disabled_s3_http_access" {
+  description = "Whether to restrict HTTP access to S3 bucket."
+  type        = bool
+  default     = true
+}