+++ aws/cloudtrail/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.3.14]() (2025-04-11)
+
+### Features
+
+* Restrict HTTP access to S3 bucket via `disabled_s3_http_access`.
+
 ## [0.1.15]() (2024-12-20)
 
 ### Features